cloth-store-api/requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.4
djangorestframework==3.13.1
emoji==1.7.0
environs==8.0.0
gunicorn==20.1.0
loguru==0.6.0
marshmallow==3.16.0
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pyparsing==3.0.9
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
